+++ build.gradle
@@ -28,12 +28,17 @@ dependencies {
     // batch
     implementation 'org.springframework.boot:spring-boot-starter-batch'
 
+    //quartz
+    implementation 'org.springframework.boot:spring-boot-starter-quartz'
+
     // mail
     implementation 'org.springframework.boot:spring-boot-starter-mail'
 
     implementation 'org.springframework.boot:spring-boot-starter-validation'
 
     implementation 'org.apache.commons:commons-lang3:3.12.0'
+
+
 
     // jwt
     implementation 'io.jsonwebtoken:jjwt:0.9.1'
