+++ src/main/java/com/base/community/BaseCommunityApplication.java
@@ -5,11 +5,13 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+import org.springframework.scheduling.annotation.EnableScheduling;
 
 @SpringBootApplication
 @EnableJpaAuditing
 @EnableJpaRepositories
 @EnableBatchProcessing
+@EnableScheduling
 public class BaseCommunityApplication {
 
     public static void main(String[] args) {
