+++ src/main/java/com/base/community/config/BatchConfig.java
@@ -9,6 +9,7 @@ import org.springframework.batch.core.configuration.annotation.JobBuilderFactory
 import org.springframework.batch.core.configuration.annotation.JobScope;
 import org.springframework.batch.core.configuration.annotation.StepBuilderFactory;
 import org.springframework.batch.repeat.RepeatStatus;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
@@ -22,16 +23,17 @@ public class BatchConfig {
 
 
     @Bean
-    public Job testJob() {
-        return jobBuilderFactory.get("testJob")
-                .start(this.testStep())
+    public Job Job() {
+        return jobBuilderFactory.get("Job")
+                .start(this.Step2())
+                .next(this.Step1())
                 .build();
     }
 
     @Bean
     @JobScope
-    public Step testStep(){
-        return stepBuilderFactory.get("testStep")
+    public Step Step1(){
+        return stepBuilderFactory.get("Step1")
                 .tasklet((contribution, chunkContext) -> {
                     jobPostingService.insertJopPosting();
                     return RepeatStatus.FINISHED;
@@ -39,4 +41,14 @@ public class BatchConfig {
                 .build();
     }
 
+    @Bean
+    @JobScope
+    public Step Step2(){
+        return  stepBuilderFactory.get("Step2")
+                .tasklet((contribution, chunkContext) -> {
+                    jobPostingService.deleteJobPosting();
+                    return RepeatStatus.FINISHED;
+                })
+                .build();
+    }
 }
